feat(analytics): separate unique movies from rewatches and enhance watchlist UX
- **DTOs**: Added `WatchedCount` property and secondary constructors to `GenreCountDto`, `DirectorCountDto`, and `ActorCountDto` to support backwards compatibility with Dapper materialization.
- **Rewatch Logic Standardization**:
  - Updated `DapperAnalyticsService` (Dashboard) to use `COUNT(DISTINCT MovieId)` for Top Genres, Directors, Actors, and Decades, ensuring rewatches don't inflate unique film counts.
  - Modified `GetTotalTimeByDirectorOrGenreAsync` in `WatchedQueriesImpl` to properly multiply runtimes by diary entry counts, preserving accurate watchtime metrics.
- **Watchlist Enhancements**: Added CTEs to `WatchlistQueriesImpl` to retrieve and join the count of unique movies already watched for directors, actors, and genres on the user's watchlist.
- **UI/UX Updates**:
  - Relabeled "Movies Watched" to "Total Watches (Inc. Rewatches)" on the Dashboard for absolute clarity.
  - Advanced Stats now visually display both `pending` and `watched` counts side-by-side for watchlist items.
  - Introduced dynamic height calculations (`getChartMaxValue`) for bar charts in `stats.ts`.
  - Added empty states ("No viewing history found...") and elegant styling for unrated items.
- **Fixes**: Casted `AverageRating` aggregations to `DOUBLE PRECISION` across Dapper queries to resolve `System.Decimal` deserialization crashes.

+++ backend/Frametric.Application/DTOs/Analytics/DecadeCountDto.cs
@@ -1,3 +1,4 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record DecadeCountDto(int Decade, long Count);
+public record DecadeCountDto(int Decade, int Count);
+
