+++ backend/Frametric.Application/DTOs/Analytics/MonthlyActivityDto.cs
@@ -1,3 +1,4 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record MonthlyActivityDto(int Month, long Count);
+public record MonthlyActivityDto(int Month, int Count);
+
