+++ backend/Frametric.Application/DTOs/Analytics/MonthlyWatchesDto.cs
@@ -1,3 +1,4 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record MonthlyWatchesDto(int Month, long Count);
+public record MonthlyWatchesDto(int Month, int Count);
+
